pyproject.toml
[project]
name = "convocation"
version = "0.1.0"
description = "ConvocAItion — Your community's site. Owned by everyone."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.14",
    "pydantic[email]>=2.10",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "argon2-cffi>=23.1",
    "python-multipart>=0.0.18",
    "httpx>=0.28",
    "gitpython>=3.1",
    "jinja2>=3.1",
    "markdown>=3.7",
    "pywebpush>=2.0",
    "py-vapid>=1.9",
    "itsdangerous>=2.2",
    "python-frontmatter>=1.1",
    "diff-match-patch>=20230430",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"
